pyproject.toml
[project]
name = "stud-iu-back"
version = "0.1.0"
description = ""
authors = [
    {name = "ekirill",email = "kirillv.2016@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.12"

dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
